// src/app/SorenTeb/model/Data/Reception/reception-model-by-id-model-dto.ts
import {ReceptionServiceModelByIdModelDto} from './reception-service-model-by-id-model-dto';

export class ReceptionModelByIdModelDto {
  id: number;
  receptionCode: string;
  documentCode: string;
  baseParaclinicChildId: number;
  paraclinicChildName: string;
  receptionDate: string;
  receptionTime: string;
  firstName: string;
  lastName: string;
  sex: number;
  sexName: string;
  age: number;
  ageType: number;
  ageTypeName: string;
  birthDate: string;
  rankNo: number;
  birthLocation: number;
  tel: string;
  mobile: string;
  fatherName: string;
  patientAddress: string;
  insuranceCode: number;
  insuranceName: string;
  insurancePercent: number;
  insuranceVersionDate: string;
  insuranceBookNumber: number;
  insurancePageNumber: number;
  insuranceExpireDate: string;
  iSDisabled: boolean;
  iSDisabledDesc: string;
  nationNumber: string;
  insuranceRate: number;
  insuranceShareRate: number;
  patientShareRate: number;
  freeRate: number;
  mustPay: number;
  outOfRate: number;
  discountRate: number;
  paymentRate: number;
  baseBeneficiaryId: number;
  beneficiaryFirstName: string;
  beneficiaryLastName: string;
  fullBeneficiaryName: string;
  beneficiaryGroupName: string;
  beneficiarySpecialName: string;
  patientStatus: number;
  description: string;
  introducation: string;
  isDM: boolean;
  isFH: boolean;
  isHTN: boolean;
  isHIP: boolean;
  isSM: boolean;
  isOther: boolean;
  queueNumber: number;
  introduction2: number;
  createUserId: number;
  updateUserId: number;
  createDate: string;
  createTime: string;
  updateDate: string;
  updateTime: string;
  receptionServiceModelByIdModel: Array<ReceptionServiceModelByIdModelDto>;
}