// src/app/SorenTeb/model/Data/Reception/reception-service-model-by-id-model-dto.ts
export class ReceptionServiceModelByIdModelDto {
  id: number;
  serviceName: string;
  receptionId: number;
  receptionCode: string;
  serviceID: number;
  indexQuantity: number;
  freeRate: number;
  insuranceRate: number;
  insuranceShareRate: number;
  patientShareRate: number;
  outOfRate: number;
  iSFree: boolean;
  iSNC: boolean;
}